[feat] Skip shp creation step
For legacy reasons, shape file can't be more than 4GB. The computation
of data needs more. To skip this limitation, shapes are just no more
created. Instead of that, postgres is directly loaded with
gdal_contour.

Beware of table structure which is a bit different : geom is replaced
by wkb_geometry !